Fail fast when a new database transaction cannot be started.
We need a new approach to handle database failures. Since db.Begin() blocks
when the database server is not responding, the current implementation, in the
worst case, sleeps a given duration for each request. This cannot be an option;
it can exhaust server resources quickly under high loads.  This commit
implements a simple solution that starts a goroutine to check database
periodically. When the database server still freezed all the following requests
will fail and will not wait sleeping.