Develops new error function to work with follow-up
In some cases the user wants to add an extra layer over the current Error type,
and on those cases we can't go back one invocation up to identify the error
caller. For that reason a new function was created in a way that is possible
for the user to define the number of invocations to follow up and find the
error caller.